Ensure keywords are trimmed when syncing translations
The lexer parses keywords with spaces following them, and then trims the
result to get only the keyword. This assumes that the keyword itself was
trimmed, otherwise the detection of the keyword type breaks (and the
tests break too).
The Tamil translations were including spaces in the translations. This
trims them during the translation update (better than trimming them each
time they are loaded, even though it would be the safer way).