Rewrite the i18n tests to parse in the test rather than the data provider
A failure in a data provider triggers a PHPUnit warning and skips all the
tests. This does not allow to properly test for cases where some languages
are triggering an error during parsing.
The test is now rewritten to perform the parsing in the test, allowing each
language test to run separately.