Provide better messages for TableNode construct errors
The 'array is not two dimensional' error message is returned when any of
the multitude construct checks fails. The message is too generic and
does not inform user about what is actually wrong with his data.

Remove one of of the duplicate checks for $row being array.

Construct specific error messages for each failure with indicating
row/column position of the error, expected invariant and what is wrong
with supplied data.