chore: Fix cleanup of test directories when running FileCacheTest
The test `tearDown` was only attempting to remove `*.feature.cache`
in the configured path, however the FileCache actually writes
files in a directory below that path. Therefore nothing was being
deleted and the files from one test were still present for the
next (and at the end of the run).

Generate a unique directory for every test and use
`symfony/filesystem` to ensure that the entire directory tree is
properly cleaned up.