change: Deprecate `Gherkin::VERSION` and change how file caches are namespaced
The `Gherkin::VERSION` constant has not been updated for recent releases,
and will not be part of our release process in the future.

This constant was only being used internally to ensure that any pre-existing
file caches are invalidated when a different version of the parser is
installed. I have switched this to get a version from the composer runtime
API, which should be at least as reliable as the previous hardcoded
constant for this purpose.