Rewrite GitHub action and script for updating Gherkin dev dependency
This has been broken since cucumber reorganised their repositories. As
well as moving to a new repository, the tagging and changelog structure
is different. There were also some issues with the old scripts - e.g.
the changelog was not properly escaped in the PR body (see #255 for an
example of broken content & links).

I have rewritten the flow so that:

* We use the GitHub API to fetch release details and changelogs, rather
  than trying to parse the changelog markdown file.
* We only bump one Gherkin release per workflow run (so that we can see
  clearly which release causes tests to fail, if any).
* We include the Gherkin release details in the commit message (the old
  script could bump multiple releases in a single commit but only showed
  the most recent changelog in the PR body, which was misleading).
* If there is already a PR open, we append subsequent releases to the
  same PR.