fix: Ensure assertions in the update_cucumber script are executed
By default shivammathur/setup-php configures PHP with a production
ini file, which disables zend assertions.

This means that calls like `assert(preg_match(...,..., $matches))`
are compiled out, leaving $matches undefined. The logic in the
update_cucumber script requires the assertions to run, so we
use the development ini file and have the script fail if
they're not enabled.