test: Ignore feature description indent when testing cucumber parity
We handle indentation of feature descriptions differently to cucumber.
Therefore, we were previously skipping the cucumber/gherkin test example
that had an indented feature description.

However, that example was primarily proving the behaviour of escapes
when parsing Gherkin tables - and there are other examples in later
Gherkin releases that we want to be able to run.

Therefore, strip all leading whitespace in feature descriptions when
testing for cucumber parity so that we can assert the rest of the parsed
example matches.