test: Ignore `extra_table_content` cucumber parity example
Cucumber's parsers ignore content to the right of the final `|` table
column separator, with the comment "It is not recommended to use this
feature, but it is how the implementation currently works." (see
https://redirect.github.com/cucumber/gherkin/pull/245 and the chain
of related PRs).

However, we have an explicit test in
`ParserExceptionsTest::testTableWithoutRightBorder` that the parser
will throw an exception if it encounters content to the right of the
closing `|`.

For now, ignore this example to be addressed as a separate issue / PR.