test: Ignore `descriptions_with_comments` cucumber parity example
cucumber/gherkin@v31.0.0 introduced the ability to have comments inside
descriptions (in https://redirect.github.com/cucumber/gherkin/pull/334).

We do not yet support this, and their test examples rely on being able
to add a description to an example table which we don't support either
(we have an open issue for this).

Add this to the exclusions from our parity tests for now.