change: Throw ParserException if file ends with tags
For parity with cucumber/gherkin, do not accept feature files that
end with `@tags` with no subsequent content.

cucumber/gherkin is relatively permissive about incomplete files in
some cases (for example, it is explicitly valid to have a file with a
Feature: but no scenarios, or a Scenario: with no steps). I suspect this
is intentional to avoid parser errors while iterating on a feature.

However, the presence of `@tags` at the end of the file is unusual and
suggests that the file is incomplete or damaged. cucumber/gherkin treats
this as an exception and so should we.