change: Don't trim description when loading cucumber NDJson files
We already know that the way we trim descriptions is not consistent
with cucumber, and are now normalising that in the CompatibilityTest.
As pointed out by @uuf6429, we were also trimming the actual value
parsed by cucumber when loading their NDJson files to test against.

Remove that to avoid causing further confusion about exactly what
the whitespace is expected to look like.