change: Throw on unexpected Steps following an Examples: table
We were not previously enforcing that an Examples: table can only be
followed by more Examples: tables, comments, whitespace or the end of
the Scenario / file.

With this commit, that will now cause an UnexpectedParserNodeException.
I have refactored the order of checking whether we can accept a
particular node when parsing a scenario / outline body to make the logic
easier.

It also slightly changes the message of the exception for existing
invalid cases so that they are more accurate about what is expected in
the current state.