Merge pull request #294 from uuf6429/bugfix/verify-optional-table-header-exists
test: Handle optional tableHeader when loading NDJson examples

Previously, when loading NDJson ASTs for gherkin compatibility
tests, we were assuming that an `examples` object would always 
have a `tableHeader` key.

In fact, this is an optional property in the schema, which is not
present if the feature file contains an `Examples:` header but
no following table rows (which for us equates to the 
`Examples:` not being present at all).

We now cope with this, and are also stricter that the example
table properties match our expected schema in case of future
changes.