add: Parse `Scenario` as a synonym for `Scenario Outline` (#316)
In cucumber/gherkin, the `Scenario:` and `Scenario Outline:` keywords
are interchangeable. Either may have (or not have) an `Examples:` table.

We still need to differentiate between `OutlineNode` and `ScenarioNode`
in the parsed result, for BC and because Behat treats these differently
at runtime.

However, we can safely parse a `Scenario:` that contains an `Examples:`
table (which would have previously been invalid) to an `OutlineNode` -
exactly as if the user had written `Scenario Outline:`.

This does require refactoring how we parse tags, because:

* Previously, we could make assumptions based on whether we were
expecting (or had already seen) an Examples table.
* Now, we can't tell from the tags alone whether they belong to an
Examples table, or mark the end of the current Scenario and the start of
the next.

Therefore, we now scan ahead to predict which node the tags are attached
to. This also allows us to throw a clearer exception if tags are found
in unexpected places in the document.

Relates to #153