test: Match expected structure when loading NDJson for incomplete tables
Our concepts of example nodes and example table nodes differ slightly
from those in cucumber/gherkin. As a result, when we recently changed
the NDJson table parsing, we slightly misunderstood how some of the
edge cases for incomplete tables should be represented in our own
object graph.

This was causing a false test failure as our parser was correctly
reading incomplete ExampleTableNode objects that we had ignored when
loading the NDJson.