change: Parse Scenario Outline without examples as a Scenario
To match the behaviour of cucumber/gherkin, we no longer throw an
exception if a feature file contains a `Scenario Outline:` with no
examples.

Instead, this will be parsed as a `Scenario` node.

Note that the presence of an `Examples:` node does not guarantee there
are actually any examples to run - it is valid to have `Examples:` with
no table, or no table rows.