test: Refactor valid ParserExceptionsTest examples into cucumber/gherkin testdata (#322)
This is a first step towards #321 and some of the cucumber-parity issues
related to when we do / do not throw an exception.

For the examples in ParserExceptionsTest where we parse (or throw) the
same as the official cucumber/gherkin parser, I have extracted the
test to the gherkin testdata format, and added it to our CompatibilityTest.

This will allow us to contribute these upstream and/or maintain our parity
examples in future.