change: Improve unexpected Lexer state exception when setting language
I have attempted to add a test for this, but it is actually not possible
to trigger it using the public interface of the class. This is due to
the sequence of logic between stashing / deferring tokens and the
`allowLanguageTag` flag.

I have therefore excluded from code coverage as this is essentially dead
code that is only present to provide extra protection for users with
custom extensions.