change: Do not throw on some unexpected Feature / Language tags (#323)
Historically, we have thrown exceptions if:

* A `Feature:` line is found after the first `Feature:` line
* A language comment is found after the first language comment.

This is not the case in cucumber/gherkin, which stops looking for
features and languages after it finds them - they are just treated as
text / comments and are valid in places where those are allowed.

I have changed our behaviour to match cucumber/gherkin.