Use a custom object comparator to ignore the keywordType of StepNode
The keywordType of our StepNode is not a concept that exists in the
Gherkin AST of cucumber. It actually corresponds to the pickle step type
in their architecture.
Using a dedicated object comparator (registered only for cucumber
compatibility tests) allows to ignore that property without having to
normalize steps, which makes the code a lot clearer.

The handling of the feature description is kept in a normalization step
as this is related to a parity difference with cucumber which is meant
to be solved at some point instead.