test: Enable gherkin test that now passes, update reason for skipping
The `incomplete_scenario` cucumber/gherkin example was excluded but
actually now passes - I have not identified which of the recent parser
changes has fixed it, but we no longer need to skip it.

`incomplete_background_2` also essentially parses correctly, other than
the background description being parsed as part of the title which we
already know about. I have updated the reason for exclusion.