change: Ignore content after table right-hand `|` (instead of throwing)
For parity with cucumber/gherkin - their `extra_table_content` testdata
notes that "It is not recommended to use this feature, but it is how the
implementation currently works."

The implementation is taken from the equivalent code in cucumber/gherkin
StringGherkinLine.

This should be BC for end-users - previously we treated table rows
that did not end with a `|` as plain text, which in turn triggered a
ParserException as there is nowhere that plain text can immediately
follow a table row.

Now these will instead be parsed, and just the right-hand "column" /
content will be ignored.