change: Throw ParserException if Background comes after first Scenario (#343)
Gherkin requires that any Feature `Background:` must come before the
first Scenario / Scenario Outline in a file (to match the execution
order of these steps).

We were previously allowing the background to appear anywhere at the top
(Feature) level of the file. This will now throw a ParserException.

Although this could theoretically cause existing features to break, I do
not consider it to be a BC break - we are now just being stricter that
the input matches the expected schema.

Fixes #328