Ensure a valid instantiation of the ArrayKeywords in tests
ArrayKeywords makes its language default to `en` and assumes that there
is data available for it without checking. When this is not the case,
calling its methods before setting another language will trigger a PHP
warning and return `null` (in methods where the phpdoc says it must
return a string).
The ArrayKeywords test is currently passing only because the current way
the parser works avoids this broken state for the particular case being
used in the test, but this is not guaranteed to stay true.