change: Deprecate extending node classes
These are DTOs representing our AST and there should be no need
for external users to extend them.

The only exception is TableNode, which cannot be final because
we extend it ourselves (for ExampleTableNode) and has logic for
accessing different representations of its content that other
users may also be extending.