change: Add ParserInterface, deprecate extending Lexer, Parser & nodes (#354)
As per #347 there is no reason for end-users to extend these classes,
and leaving them open for extension complicates maintenance. The
ParserInterface provides for users who wish to mock (or entirely
replace) the Parser.

They will be marked as `final` for the 5.0.0 release.

The only Node class I have not marked as `@final` is TableNode - we
extend it ourselves for ExampleTableNode, and it contains more logic
than a simple DTO. This should be considered separately for 5.0.

Fixes #347