fix: Use default file cache key if `behat/gherkin` version is unknown (#370)
As per PHPStan, it is possible that the version of the `behat/gherkin`
package, (which we use as cache key), cannot be retrieved. This
would cause a php deprecation due to attempting to hash a null
value.

This PR makes it default to `'unknown'` in that case.

This change was extracted from #364.