ci: Fix error in cucumber update job due to non-autoloadable classes
#365 extracted some filesystem logic from these scripts to the new
Filesystem helper class. This broke the job, because the scripts are
unable to locate the class. This is because there is no autoloader
registered in this environment (and, in the case of the main
update_cucumber script, we have not yet run composer install).

Fixed by manually requiring the class for update_cucumber, and
registering the autoloader for the update_i18n (which runs after the
composer install).