ci: Fix incorrect assumption about github releases in update_cucumber (#380)
#368 converted the previous `readJsonFileArray` utility method to a
stricter `readJsonFileHash` to improve phpstan checking. This was based
on the belief that we were always currently using that method to parse
JSON data with an object structure at the root.

However, in the case of the update_cucumber script, the github releases
API returns a JSON array of objects. The strict check was therefore
causing that job to fail.

I have introduced the looser `readJsonFileArray` for this purpose and
confirmed that the task now works locally.