test: Gherkin with an invalid language tag is parsed as `en`
Prior to the implementation of the new DialectProvider, if a gherkin
file contained a language tag for an unknown language then the Lexer and
Parser would fall back to parsing in `en` BUT the FeatureNode->language
would carry the original value from the file.

This has now changed - the FeatureNode always contains the language
that was actually parsed (because it comes from the Lexer's
`currentDialect`, which defaults to `en` and is only overwritten if the
dialect exists).