test: Explicitly cover departures from expected cucumber parsing (#392)
Previously, we were just skipping testing gherkin examples where our
parsing differed from Cucumber. In some cases, these contain syntax that
is not represented in our own test data.

This created a risk that, as we introduce the new compatibility mode, we
accidentally change the behaviour of the parser in legacy mode.

Going forward, we will explicitly assert that our parsing in these cases
matches our existing behaviour.

I have implemented this by capturing and dumping the result to YAML,
rather than by manually representing in e.g. ndjson and loading this to
compare as a FeatureNode.

This is because our existing test logic, loaders and comparators are
intentionally "loose" in certain places where we expect the serialised
expectation to differ from the runtime value (some string padding,
exception messages, etc) whereas we want to be absolutely explicit about
the content of our nodes.

Serialising to YAML and comparing as a string will also make it easy to
update the expectations (by running with RE_RECORD_EXPECTATIONS=1) when
e.g. we add new properties to our Node classes, or intentionally change
exception messages.

The expectations in this branch are based off the 4.14.0 release, before 
we introduced any of the new compatibility mode parser changes, to 
ensure that the expectations reflect the behaviour of the current 
production release.