fix: update_cucumber script should not fail on manually created releases
The script was assuming that `target_commitish` in the releases API
would generally be the hash of the release tag, other than for an edge
case on v29.0.0 which had been patched manually in the code.

In fact, `target_commitish` will only be a hash if the tag was created
before the release (e.g. by cucumber's automated release actions). If
the release is created manually in the github UI with a not-yet-existing
tag then `target_commitish` will instead be the branch name it was
created from. In that case, the only way to get the hash is to
separately look up the tag name in the tags API.

This was causing the script to fail due to two recent manual releases on
the upstream repo.

I have fixed this to now look up tag details if they're missing from the
releases API. As part of this, I have refactored to only look up
releases that are `>=` the one we are currently pinned to. This means we
will only need to look up tags for manual releases until we update past
them.