feat: Support parsing all describable nodes in `gherkin-32` mode (#361)
Historically, we have not supported parsing descriptions as a separate
concept for Examples, Background, Scenario or Scenario Outline. Text
following the keyword line was instead parsed as a multi-line title.

This PR implements support for instead capturing these as standalone
properties when running in cucumber/gherkin compatibility mode.

This borrowed from and builds on work done by @jojo1981 in #254, 
but started from a clean branch as there have been a lot of changes
since that PR was opened.

Fixes #154 #211