fix: Throw on tag whitespace in `gherkin-32` compatibility mode
When parsing tags, we (and cucumber/gherkin) split on the `@` character.
So `@some tag @values` will parse as `['@some tag', '@value']`. This may
be unexpected if e.g. a user has simply missed a `@` character.

Therefore cucumber/gherkin will throw if a tag contains any whitespace.
Historically we have accepted this, although we have triggered a
deprecation since v4.9.0.

Update the parser to throw if in `gherkin-32` compatibility mode.